third_party/WebKit/LayoutTests/inspector/console/console-time-expected.txt

console.time / console.timeEnd tests.

console-time.html:12 default: <time>
console-time.html:14 42: <time>
console-time.html:16 239: <time>
console-time.html:18 [object Object]: <time>
